Move UAPI to new module
