+++ wg-traits/src/lib.rs
@@ -1,6 +1,8 @@
+mod configuration;
 mod endpoint;
 pub mod tun;
 pub mod uapi;
 pub mod udp;
 
+pub use configuration::Configuration;
 pub use endpoint::Endpoint;
