+++ wg-uapi/src/uapi/error.rs
