+++ wireguard-rs/src/configuration/config.rs
@@ -3,32 +3,16 @@ use std::sync::atomic::Ordering;
 use std::time::{Duration, SystemTime};
 
 use wg_crypto::PSK;
-use wg_traits::udp::Owner;
-use wg_traits::{Endpoint, tun, udp};
+use wg_traits::{
+    Configuration, Endpoint, tun,
+    udp::{self, Owner},
+};
+use wg_uapi::uapi::{ConfigError, PeerState};
 use x25519_dalek::{PublicKey, StaticSecret};
 
-use super::*;
+use crate::wireguard::WireGuard;
 
 const PROTOCOL_VERSION: usize = 1;
-
-/// The goal of the configuration interface is, among others,
-/// to hide the IO implementations (over which the WG device is generic),
-/// from the configuration and UAPI code.
-///
-/// Furthermore it forms the simpler interface for embedding WireGuard in other applications,
-/// and hides the complex types of the implementation from the host application.
-///
-/// Describes a snapshot of the state of a peer
-pub struct PeerState {
-    pub rx_bytes: u64,
-    pub tx_bytes: u64,
-    pub last_handshake_time: Option<(u64, u64)>,
-    pub public_key: PublicKey,
-    pub allowed_ips: Vec<(IpAddr, u32)>,
-    pub endpoint: Option<SocketAddr>,
-    pub persistent_keepalive_interval: u64,
-    pub preshared_key: PSK,
-}
 
 pub struct WireGuardConfig<'device, T: tun::Tun, B: udp::PlatformUDP> {
     wireguard: &'device WireGuard<T, B>,
@@ -49,137 +33,6 @@ impl<'device, T: tun::Tun, B: udp::PlatformUDP> WireGuardConfig<'device, T, B> {
 }
 
 /// Exposed configuration interface
-pub trait Configuration {
-    fn up(&mut self, mtu: usize) -> Result<(), ConfigError>;
-
-    fn down(&mut self);
-
-    /// Updates the private key of the device
-    ///
-    /// # Arguments
-    ///
-    /// - `sk`: The new private key (or None, if the private key should be cleared)
-    fn set_private_key(&self, sk: Option<StaticSecret>);
-
-    /// Returns the private key of the device
-    ///
-    /// # Returns
-    ///
-    /// The private if set, otherwise None.
-    fn get_private_key(&self) -> Option<StaticSecret>;
-
-    /// Returns the protocol version of the device
-    ///
-    /// # Returns
-    ///
-    /// An integer indicating the protocol version
-    fn get_protocol_version(&self) -> usize;
-
-    fn set_listen_port(&mut self, port: u16) -> Result<(), ConfigError>;
-
-    /// Set the firewall mark (or similar, depending on platform)
-    ///
-    /// # Arguments
-    ///
-    /// - `mark`: The fwmark value
-    ///
-    /// # Returns
-    ///
-    /// An error if this operation is not supported by the underlying
-    /// "bind" implementation.
-    fn set_fwmark(&mut self, mark: Option<u32>) -> Result<(), ConfigError>;
-
-    /// Removes all peers from the device
-    fn _replace_peers(&self);
-
-    /// Remove the peer from the
-    ///
-    /// # Arguments
-    ///
-    /// - `peer`: The public key of the peer to remove
-    ///
-    /// # Returns
-    ///
-    /// If the peer does not exists this operation is a noop
-    fn remove_peer(&self, peer: &PublicKey);
-
-    /// Adds a new peer to the device
-    ///
-    /// # Arguments
-    ///
-    /// - `peer`: The public key of the peer to add
-    ///
-    /// # Returns
-    ///
-    /// A bool indicating if the peer was added.
-    ///
-    /// If the peer already exists this operation is a noop
-    fn add_peer(&self, peer: &PublicKey) -> bool;
-
-    /// Update the psk of a peer
-    ///
-    /// # Arguments
-    ///
-    /// - `peer`: The public key of the peer
-    /// - `psk`: The new psk or None if the psk should be unset
-    ///
-    /// # Returns
-    ///
-    /// An error if no such peer exists
-    fn set_preshared_key(&self, peer: &PublicKey, psk: [u8; 32]);
-
-    /// Update the endpoint of the
-    ///
-    /// # Arguments
-    ///
-    /// - `peer': The public key of the peer
-    /// - `psk`
-    fn set_endpoint(&self, peer: &PublicKey, addr: SocketAddr);
-
-    /// Update the endpoint of the
-    ///
-    /// # Arguments
-    ///
-    /// - `peer': The public key of the peer
-    /// - `psk`
-    fn set_persistent_keepalive_interval(&self, peer: &PublicKey, secs: u64);
-
-    /// Remove all allowed IPs from the peer
-    ///
-    /// # Arguments
-    ///
-    /// - `peer': The public key of the peer
-    ///
-    /// # Returns
-    ///
-    /// An error if no such peer exists
-    fn _replace_allowed_ips(&self, peer: &PublicKey);
-
-    /// Add a new allowed subnet to the peer
-    ///
-    /// # Arguments
-    ///
-    /// - `peer`: The public key of the peer
-    /// - `ip`: Subnet mask
-    /// - `masklen`:
-    ///
-    /// # Returns
-    ///
-    /// An error if the peer does not exist
-    fn add_allowed_ip(&self, peer: &PublicKey, ip: IpAddr, masklen: u32);
-
-    fn get_listen_port(&self) -> Option<u16>;
-
-    /// Returns the state of all peers
-    ///
-    /// # Returns
-    ///
-    /// A list of structures describing the state of each peer
-    fn get_peers(&self) -> Vec<PeerState>;
-
-    fn get_fwmark(&self) -> Option<u32>;
-}
-
 fn start_listener<T: tun::Tun, B: udp::PlatformUDP>(
     cfg: &mut WireGuardConfig<T, B>,
 ) -> Result<(), ConfigError> {
@@ -209,7 +62,10 @@ fn start_listener<T: tun::Tun, B: udp::PlatformUDP>(
     Ok(())
 }
 
-impl<'device, T: tun::Tun, B: udp::PlatformUDP> Configuration for WireGuardConfig<'device, T, B> {
+impl<'device, T: tun::Tun, B: udp::PlatformUDP>
+    Configuration<ConfigError, PeerState, PublicKey, StaticSecret>
+    for WireGuardConfig<'device, T, B>
+{
     fn up(&mut self, mtu: usize) -> Result<(), ConfigError> {
         log::info!("configuration, set device up");
         self.wireguard.up(mtu);
